Rename data repository submodule.
