+++ coxeter/shape_families/__init__.py
@@ -3,7 +3,7 @@ from .plane_shape_families import Family332, Family432, Family532
 from .tabulated_shape_family import (TabulatedShapeFamily,
                                      TabulatedGSDShapeFamily)
 from .common_families import RegularNGonFamily, PlatonicFamily
-from .data_repositories import family_from_doi
+from .doi_data_repositories import family_from_doi
 
 __all__ = ['ShapeFamily', 'TabulatedShapeFamily', 'TabulatedGSDShapeFamily',
            'shape_repositories', 'Family332', 'Family432', 'Family532',
